# docs/spec/openapi.yml
openapi: 3.0.2
servers:
  - url: /v1
info:
  version: 1.0.0
  title: Percona Everest schema
  description: |
    # Authentication
    All requests to Everest API require `Authorization: Bearer <token>` header with a valid token in plain-text.
    
    The token can be obtained by using `everestctl token reset` which resets the token and prints it to the screen.
...
paths:
...
  '/monitoring-instances':
    post:
      tags:
        - monitoringInstances
      summary: Create a new monitoring instance object
      description: |
        A monitoring instance object requires `type` to be set.
        Based on the `type` the respective key with configuration needs to be set.
        Such as, if `type: pmm`, then `pmm` key needs to be provided with a configuration.
      operationId: createMonitoringInstance
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MonitoringInstance'
        '400':
          description: Unsuccessful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        description: The Monitoring instance object to be created
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MonitoringInstanceCreateParams'
    get:
      tags:
        - monitoringInstances
      summary: List of the created monitoring instances
      description: List of the created monitoring instances
      operationId: listMonitoringInstances
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MonitoringInstancesList'
        '400':
          description: Unsuccessful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  '/monitoring-instances/{name}':
    get:
      tags:
        - monitoringInstances
      summary: Get the specified monitoring instance
      description: Get the specified monitoring instance
      operationId: getMonitoringInstance
      parameters:
        - name: name
          in: path
          description: Name of the Monitoring instance
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MonitoringInstance'
        '400':
          description: Unsuccessful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Monitoring instance not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      tags:
        - monitoringInstances
      summary: Update the specified Monitoring instance
      description: Update the specified Monitoring instance
      operationId: updateMonitoringInstance
      parameters:
        - name: name
          in: path
          description: Name of the monitoring instance
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MonitoringInstance'
        '400':
          description: Unsuccessful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Monitoring instance not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        description: The monitoring instance object to be updated.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MonitoringInstanceUpdateParams'
    delete:
      tags:
        - monitoringInstances
      summary: Delete the specified Monitoring instance
      description: Delete the specified Monitoring instance
      operationId: deleteMonitoringInstance
      parameters:
        - name: name
          in: path
          description: Name of the monitoring instance
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Successful operation
        '400':
          description: Unsuccessful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Monitoring instance not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    Error:
      type: object
      description: Error response
      properties:
        message:
          type: string
...
    MonitoringInstanceBase:
      type: object
      description: Monitoring instance information
      properties:
        type:
          type: string
          enum:
          - pmm
          x-go-type-skip-optional-pointer: true
        url:
          type: string
          minLength: 1
          x-go-type-skip-optional-pointer: true
        targetNamespaces:
          type: array
          description: List of namespaces allowed to use this monitoring instance
          items:
            type: string
    MonitoringInstanceBaseWithName:
      type: object
      description: Monitoring instance information
      allOf:
        - $ref: '#/components/schemas/MonitoringInstanceBase'
        - type: object
          properties:
            name:
              type: string
              minLength: 1
              description: A user defined string name of the storage in the DNS name format https://kubernetes.io/docs/concepts/overview/working-with-objects/names/#dns-label-names
              x-go-type-skip-optional-pointer: true
    MonitoringInstancePMM:
      type: object
      properties:
        pmm:
          type: object
          x-go-type-name: PMMMonitoringInstanceSpec
          properties:
            user:
              type: string
              x-go-type-skip-optional-pointer: true
            password:
              type: string
              x-go-type-skip-optional-pointer: true
            apiKey:
              type: string
              minLength: 1
              x-go-type-skip-optional-pointer: true
    MonitoringInstanceCreateParams:
      description: Monitoring instance create information
      allOf:
        - $ref: '#/components/schemas/MonitoringInstanceBaseWithName'
        - $ref: '#/components/schemas/MonitoringInstancePMM'
      required:
        - type
        - url
        - targetNamespaces
        - name
    MonitoringInstanceUpdateParams:
      description: Monitoring instance update information
      allOf:
        - $ref: '#/components/schemas/MonitoringInstanceBase'
        - $ref: '#/components/schemas/MonitoringInstancePMM'
    MonitoringInstance:
      description: Monitoring instance information
      allOf:
        - $ref: '#/components/schemas/MonitoringInstanceBaseWithName'
      required:
        - type
        - url
        - targetNamespaces
        - name
    MonitoringInstancesList:
      type: array
      items:
        type: object
        $ref: '#/components/schemas/MonitoringInstance'
...